requirements.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
distlib==0.3.8
dj-rest-auth==6.0.0
Django==5.0.2
django-allauth==0.62.1
django-cors-headers==4.3.1
django-oauth2-provider==0.2.6.1
django-rest-auth==0.9.5
django-tenant-schemas==1.12.0
django-tenants==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.4
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
oauth2_provider==0.0
oauthlib==3.2.2
ordered-set==4.1.0
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==69.5.1
shortuuid==1.0.13
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
twilio==9.0.4
tzdata==2023.4
urllib3==2.2.1
virtualenv==20.25.1
yarl==1.9.4
